Support from ghc-8.6 to ghc-9.0
