+++ plugins/hls-gadt-plugin/test/testdata/Infix.expected.hs
@@ -2,4 +2,4 @@ module Infix where
 
 data Foo where
   (:->) :: Int -> Char -> Foo
-  deriving (Show)
+  deriving ()
