+++ plugins/hls-gadt-plugin/test/testdata/SingleDeriving.expected.hs
@@ -1,0 +1,5 @@
+module SingleDeriving where
+
+data Foo a b where
+  Bar :: b -> a -> Foo a b
+  deriving Eq
