+++ plugins/hls-gadt-plugin/test/testdata/SingleDeriving.hs
@@ -1,0 +1,4 @@
+module SingleDeriving where
+
+data Foo a b = Bar b a
+ deriving (Eq)
