+++ plugins/hls-gadt-plugin/test/testdata/SingleDerivingGHC92.expected.hs
@@ -1,0 +1,5 @@
+module SingleDerivingGHC92 where
+
+data Foo a b where
+  Bar :: b -> a -> Foo a b
+  deriving (Eq)
