+++ plugins/hls-gadt-plugin/test/testdata/SingleDerivingGHC92.hs
@@ -1,0 +1,4 @@
+module SingleDerivingGHC92 where
+
+data Foo a b = Bar b a
+ deriving (Eq)
